created new class for common logic
