+++ src/main/java/de/dennisguse/opentracks/data/models/TrackPoint.java
@@ -14,6 +14,7 @@
  * the License.
  */
 package de.dennisguse.opentracks.data.models;
+import de.dennisguse.opentracks.data.Id;
 
 import android.location.Location;
 import android.os.Parcel;
@@ -468,28 +469,5 @@ public class TrackPoint {
                 ", altitudeGain_m=" + altitudeGain_m +
                 ", altitudeLoss_m=" + altitudeLoss_m +
                 '}';
-    }
-
-    public record Id(long id) implements Parcelable {
-
-        @Override
-        public int describeContents() {
-            return 0;
-        }
-
-        @Override
-        public void writeToParcel(Parcel parcel, int i) {
-            parcel.writeLong(id);
-        }
-
-        public static final Creator<Id> CREATOR = new Creator<>() {
-            public Id createFromParcel(Parcel in) {
-                return new Id(in.readLong());
-            }
-
-            public Id[] newArray(int size) {
-                return new Id[size];
-            }
-        };
     }
 }
